fix: use `uint64` instead of `uint32` for packet sequence (#361)

+++ contracts/interfaces/IICS20Transfer.sol
@@ -12,7 +12,7 @@ interface IICS20Transfer {
     /// @notice Send a transfer by constructing a message and calling IICS26Router.sendPacket
     /// @param msg_ The message for sending a transfer
     /// @return sequence The sequence number of the packet created
-    function sendTransfer(IICS20TransferMsgs.SendTransferMsg calldata msg_) external returns (uint32 sequence);
+    function sendTransfer(IICS20TransferMsgs.SendTransferMsg calldata msg_) external returns (uint64 sequence);
 
     /// @notice Send a permit2 transfer by constructing a message and calling IICS26Router.sendPacket
     /// @param msg_ The message for sending a transfer
@@ -25,7 +25,7 @@ interface IICS20Transfer {
         bytes calldata signature
     )
         external
-        returns (uint32 sequence);
+        returns (uint64 sequence);
 
     /// @notice Send a transfer by constructing a message and calling IICS26Router.sendPacket with the provided sender
     /// @dev This is a permissioned function requiring the `DELEGATE_SENDER_ROLE`
@@ -38,7 +38,7 @@ interface IICS20Transfer {
         address sender
     )
         external
-        returns (uint32 sequence);
+        returns (uint64 sequence);
 
     /// @notice Retrieve the escrow contract address
     /// @param clientId The client identifier
